+++ FTMobileSDK/FTSDKCore/BaseUtils/Base/FTConstants.m
@@ -150,7 +150,7 @@ NSString * const FT_NETWORK_TRACEPARENT_KEY = @"traceparent";
 NSString * const FT_NETWORK_ZIPKIN_SINGLE_KEY = @"b3";
 
 NSUInteger const FT_LOGGING_CONTENT_SIZE = 30720;
-NSUInteger const FT_DB_CONTENT_MAX_COUNT = 5000;
+int const FT_DB_CONTENT_MAX_COUNT = 5000;
 NSString * const FT_SCRIPT_MESSAGE_HANDLER_NAME = @"ftMobileSdk";
 
 NSUInteger const MXRMonitorRunloopOneStandstillMillisecond = 1000;
